:recycle: refactor(Server): Move folder-based project initialization logic outside the repository
